\name{data_miss}
\docType{data}
\alias{data_miss}
\title{Example data set containing missing values}
\description{
This data set is a subset of the COPDGene study data using the
    Metabolon Global Metabolomics platform. We impose missing values using
    the Mixed missingness algorithm developed by Styczynski et al. Where the
    parameters alpha, beta, and gamma were chosen to be 30, 70, and 40 percent,
    respectively.
}


\references{

Emwas AH, Roy R, McKay RT, et al. NMR Spectroscopy for Metabolomics Research.
\emph{Metabolites}. 2019;9(7).

Lee JY, Styczynski MP. NS-kNN: a modified k-nearest neighbors approach for
imputing metabolomics data. \emph{Metabolomics}. 2018;14(12):153.
}
