\name{untargeted_LCMS_data}
\docType{data}
\alias{untargeted_LCMS_data}
\title{Example data set containing missing values}
\description{
This data set is randomly generated. We impose 30 percent missing values using
    the Mixed missingness algorithm developed by Styczynski et al. Where the
    parameters alpha, beta, and gamma were chosen to be 30, 70, and 40 percent,
    respectively.
}


\references{
Lee JY, Styczynski MP. NS-kNN: a modified k-nearest neighbors approach for
imputing metabolomics data. \emph{Metabolomics}. 2018;14(12):153.
}
